move corresponding unit tests classes
